pyproject.toml
[project]
name = "sntsn-tura-power-outage-bot"
version = "0.1.0"
description = "Telegram bot for power outage alerts in SNTSN Tura"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiohttp[speedups]>=3.11.18",
    "beautifulsoup4>=4.13.4",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=5.2.1",
]
